requirements.txt
# Core dependencies
pydantic>=2.0.0,<3.0.0
pyyaml>=6.0.0
numpy>=1.24.0

# OpenAI-compatible client
openai>=1.0.0

# Environment dependencies (if not already included)
dataclasses-json>=0.6.0

# DQN Training (Required for train_dqn.py and test_pt_model.py)
torch>=2.0.0

# Development/testing (optional)
pytest>=7.4.0
black>=23.0.0
